+++ json_serializable/lib/src/helper_core.dart
@@ -38,7 +38,7 @@ abstract class HelperCore {
       escapeDartString(nameAccess(field));
 
   @protected
-  String get prefix => '_\$${element.name.nonPrivate}';
+  String get prefix => '_\$${element.name!.nonPrivate}';
 
   /// Returns a [String] representing the type arguments that exist on
   /// [element].
@@ -61,34 +61,38 @@ InvalidGenerationSourceError createInvalidGenerationError(
   FieldElement field,
   UnsupportedTypeError error,
 ) {
-  var message = 'Could not generate `$targetMember` code for `${field.name}`';
+  var message = 'Could not generate `$targetMember` code for `${field.name!}`';
 
   String? todo;
   if (error.type is TypeParameterType) {
-    message = '$message because of type '
+    message =
+        '$message because of type '
         '`${error.type.toStringNonNullable()}` '
         '(type parameter)';
 
-    todo = '''
+    todo =
+        '''
 To support type parameters (generic types) you can:
 $converterOrKeyInstructions
 * Set `JsonSerializable.genericArgumentFactories` to `true`
   https://pub.dev/documentation/json_annotation/latest/json_annotation/JsonSerializable/genericArgumentFactories.html''';
   } else if (field.type != error.type) {
-    message = '$message because of type `${typeToCode(error.type)}`';
+    try {
+      message = '$message because of type `${typeToCode(error.type)}`';
+      // ignore: avoid_catching_errors
+    } on UnimplementedError catch (ex) {
+      message = '$message because type is unimplemented ($ex)';
+    }
   } else {
     final element = error.type.element?.name;
-    todo = '''
+    todo =
+        '''
 To support the type `${element ?? error.type}` you can:
 $converterOrKeyInstructions''';
   }
 
   return InvalidGenerationSourceError(
-    [
-      '$message.',
-      if (error.reason != null) error.reason,
-      if (todo != null) todo,
-    ].join('\n'),
+    ['$message.', if (error.reason != null) error.reason, ?todo].join('\n'),
     element: field,
   );
 }
@@ -119,13 +123,15 @@ String genericClassArguments(ClassElement element, bool? withConstraints) {
   if (withConstraints == null || element.typeParameters.isEmpty) {
     return '';
   }
-  final values = element.typeParameters.map((t) {
-    if (withConstraints && t.bound != null) {
-      final boundCode = typeToCode(t.bound!);
-      return '${t.name} extends $boundCode';
-    } else {
-      return t.name;
-    }
-  }).join(', ');
+  final values = element.typeParameters
+      .map((t) {
+        if (withConstraints && t.bound != null) {
+          final boundCode = typeToCode(t.bound!);
+          return '${t.name!} extends $boundCode';
+        } else {
+          return t.name!;
+        }
+      })
+      .join(', ');
   return '<$values>';
 }