+++ json_serializable/test/annotation_version_test.dart
@@ -7,6 +7,7 @@
 @Timeout.factor(2)
 library;
 
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:json_serializable/src/check_dependencies.dart';
@@ -32,11 +33,12 @@ void main() {
 
   group('language version', () {
     test('is less than required', () async {
-      const sdkLowerBound = '2.12.0';
+      const sdkLowerBound = '3.7.0';
       await _structurePackage(
         environment: const {'sdk': '^$sdkLowerBound'},
         dependencies: {'json_annotation': _annotationLowerBound},
-        message: '''
+        message:
+            '''
 The language version ($sdkLowerBound) of this package ($_testPkgName) does not match the required range `$supportedLanguageConstraint`.
 
 Edit pubspec.yaml to include an SDK constraint of at least $supportedLanguageConstraint.
@@ -47,87 +49,66 @@ environment:
       );
     });
 
-    test(
-      'is at least the required `$supportedLanguageConstraint`',
-      () async => await _structurePackage(
+    test('is at least the required `$supportedLanguageConstraint`', () async {
+      await _structurePackage(
         dependencies: {'json_annotation': _annotationLowerBound},
         message: null,
-      ),
-    );
+      );
+    });
   });
 
-  test(
-    'missing dependency in production code',
-    () => _structurePackage(
-      message: _missingProductionDep,
-    ),
-  );
+  test('missing dependency in production code', () async {
+    await _structurePackage(message: _missingProductionDep);
+  });
 
-  test(
-    'missing dependency in example code',
-    () => _structurePackage(
+  test('missing dependency in example code', () async {
+    await _structurePackage(
       sourceDirectory: 'example',
       message:
           'You are missing a required dependency on json_annotation with a '
           'lower bound of at least "$_annotationLowerBound".',
-    ),
-  );
+    );
+  });
 
-  test(
-    'dev dependency with a production usage',
-    () => _structurePackage(
+  test('dev dependency with a production usage', () async {
+    await _structurePackage(
       devDependencies: {'json_annotation': _annotationLowerBound},
       message: _missingProductionDep,
-    ),
-  );
+    );
+  });
 
-  test(
-    'dependency with `null` constraint',
-    () => _structurePackage(
+  test('dependency with `null` constraint', () async {
+    await _structurePackage(
       dependencies: {'json_annotation': null},
       message:
           'The version constraint "any" on json_annotation allows versions '
           'before $_annotationLowerBound which is not allowed.',
-    ),
-  );
+    );
+  });
 
-  test(
-    'dependency with "any" constraint',
-    () => _structurePackage(
+  test('dependency with "any" constraint', () async {
+    await _structurePackage(
       dependencies: {'json_annotation': 'any'},
       message:
           'The version constraint "any" on json_annotation allows versions '
           'before $_annotationLowerBound which is not allowed.',
-    ),
-  );
+    );
+  });
 
-  test(
-    'dependency with too low version range',
-    () => _structurePackage(
+  test('dependency with too low version range', () async {
+    await _structurePackage(
       dependencies: {'json_annotation': '^4.0.0'},
       message:
           'The version constraint "^4.0.0" on json_annotation allows versions '
           'before $_annotationLowerBound which is not allowed.',
-    ),
-  );
+    );
+  });
 }
 
 final _jsonSerialPubspec = Pubspec.parse(
   File('pubspec.yaml').readAsStringSync(),
   sourceUrl: Uri.file('pubspec.yaml'),
 );
-
-String _fixPath(String path) {
-  if (p.isAbsolute(path)) return path;
-
-  return p.canonicalize(p.join(p.current, path));
-}
-
-final _jsonSerialPathDependencyOverrides = {
-  for (var entry in _jsonSerialPubspec.dependencyOverrides.entries)
-    if (entry.value is PathDependency)
-      entry.key: {'path': _fixPath((entry.value as PathDependency).path)}
-};
 
 final _annotationLowerBound = requiredJsonAnnotationMinVersion.toString();
 
@@ -145,53 +126,46 @@ Future<void> _structurePackage({
   Map<String, dynamic> dependencies = const {},
   Map<String, dynamic> devDependencies = const {},
 }) async {
-  final pubspec = loudEncode(
-    {
-      'name': _testPkgName,
-      'environment': environment,
-      'dependencies': dependencies,
-      'dev_dependencies': {
-        ...devDependencies,
-        'build_runner': 'any',
-        'json_serializable': {'path': p.current},
-      },
-      'dependency_overrides': _jsonSerialPathDependencyOverrides,
+  final pubspec = loudEncode({
+    'name': _testPkgName,
+    'environment': environment,
+    'dependencies': dependencies,
+    'dev_dependencies': {
+      ...devDependencies,
+      'build_runner': 'any',
+      'json_serializable': {'path': p.current},
     },
-  );
+    'dependency_overrides': {
+      'json_annotation': {
+        'path': p.canonicalize(p.join(p.current, '../json_annotation')),
+      },
+    },
+  });
 
   await d.file('pubspec.yaml', pubspec).create();
 
   /// A file in the lib directory without JsonSerializable should do nothing!
-  await d.dir(
-    'lib',
-    [
-      d.file('no_op.dart', '''
+  await d.dir('lib', [
+    d.file('no_op.dart', '''
 class NoOp {}
-''')
-    ],
-  ).create();
+'''),
+  ]).create();
 
-  await d.dir(
-    sourceDirectory,
-    [
-      d.file(
-        'sample.dart',
-        '''
+  await d.dir(sourceDirectory, [
+    d.file('sample.dart', '''
 import 'package:json_annotation/json_annotation.dart';
 
 part 'sample.g.dart';
 
 @JsonSerializable()
 class SomeClass{}
-''',
-      )
-    ],
-  ).create();
-  final proc = await TestProcess.start(
-    Platform.resolvedExecutable,
-    ['run', 'build_runner', 'build'],
-    workingDirectory: d.sandbox,
-  );
+'''),
+  ]).create();
+  final proc = await TestProcess.start(Platform.resolvedExecutable, [
+    'run',
+    'build_runner',
+    'build',
+  ], workingDirectory: d.sandbox);
 
   final lines = StringBuffer();
   await for (var line in proc.stdoutStream()) {
@@ -200,8 +174,16 @@ class SomeClass{}
   }
 
   final output = lines.toString();
-  final expectedWarningCount = message == null ? 0 : 1;
-  final warningCount = '[WARNING]'.allMatches(output).length;
+  var expectedWarningCount = message == null ? 0 : 1;
+
+  // If pkg:analyzer is out of sync with the current SDK, we can get a warning
+  // about the "SDK language version". In that case, just expect one more
+  // warning and proceed.
+  if (output.contains('W SDK language version ')) {
+    expectedWarningCount++;
+  }
+
+  final warningCount = _warningStartOfLine.allMatches(output).length;
   expect(
     warningCount,
     expectedWarningCount,
@@ -211,10 +193,15 @@ class SomeClass{}
   );
 
   if (message != null) {
-    expect(output, contains('''
-[WARNING] json_serializable on $sourceDirectory/sample.dart:
-$message'''));
+    expect(
+      output,
+      contains('''
+W json_serializable on $sourceDirectory/sample.dart:
+${LineSplitter.split(message).map((line) => '  $line').join('\n')}'''),
+    );
   }
 
   await proc.shouldExit(0);
 }
+
+final _warningStartOfLine = RegExp(r'^W ', multiLine: true);