+++ json_serializable/test/src/_json_serializable_test_input.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// @dart=3.6
+// @dart=3.8
 
 import 'dart:collection';
 
@@ -13,6 +13,7 @@ part 'checked_test_input.dart';
 part 'constants_copy.dart';
 part 'core_subclass_type_input.dart';
 part 'default_value_input.dart';
+part 'extends_jsonkey_override.dart';
 part 'field_namer_input.dart';
 part 'generic_test_input.dart';
 part 'inheritance_test_input.dart';
@@ -34,15 +35,13 @@ enum UnsupportedEnum { not, valid }
 @JsonSerializable() // ignore: invalid_annotation_target
 Object annotatedMethod() => throw UnimplementedError();
 
-@ShouldGenerate(
-  r'''
+@ShouldGenerate(r'''
 OnlyStaticMembers _$OnlyStaticMembersFromJson(Map<String, dynamic> json) =>
     OnlyStaticMembers();
 
 Map<String, dynamic> _$OnlyStaticMembersToJson(OnlyStaticMembers instance) =>
     <String, dynamic>{};
-''',
-)
+''')
 @JsonSerializable()
 class OnlyStaticMembers {
   // To ensure static members are not considered for serialization.
@@ -113,20 +112,16 @@ class GeneralTestClass2 {
   late DateTime dateOfBirth;
 
   GeneralTestClass2(this.height, String firstName, [this.lastName])
-      :
-        // ignore: prefer_initializing_formals
-        firstName = firstName;
+    // ignore: prefer_initializing_formals
+    : firstName = firstName;
 }
 
 @ShouldGenerate(r'''
-FinalFields _$FinalFieldsFromJson(Map<String, dynamic> json) => FinalFields(
-      (json['a'] as num).toInt(),
-    );
+FinalFields _$FinalFieldsFromJson(Map<String, dynamic> json) =>
+    FinalFields((json['a'] as num).toInt());
 
 Map<String, dynamic> _$FinalFieldsToJson(FinalFields instance) =>
-    <String, dynamic>{
-      'a': instance.a,
-    };
+    <String, dynamic>{'a': instance.a};
 ''')
 @JsonSerializable()
 class FinalFields {
@@ -137,17 +132,15 @@ class FinalFields {
   FinalFields(this.a);
 }
 
-@ShouldGenerate(
-  r'''
+@ShouldGenerate(r'''
 FinalFieldsNotSetInCtor _$FinalFieldsNotSetInCtorFromJson(
-        Map<String, dynamic> json) =>
-    FinalFieldsNotSetInCtor();
+  Map<String, dynamic> json,
+) => FinalFieldsNotSetInCtor();
 
 Map<String, dynamic> _$FinalFieldsNotSetInCtorToJson(
-        FinalFieldsNotSetInCtor instance) =>
-    <String, dynamic>{};
-''',
-)
+  FinalFieldsNotSetInCtor instance,
+) => <String, dynamic>{};
+''')
 @JsonSerializable()
 class FinalFieldsNotSetInCtor {
   final int a = 1;
@@ -157,13 +150,11 @@ class FinalFieldsNotSetInCtor {
 
 @ShouldGenerate(r'''
 SetSupport _$SetSupportFromJson(Map<String, dynamic> json) => SetSupport(
-      (json['values'] as List<dynamic>).map((e) => (e as num).toInt()).toSet(),
-    );
+  (json['values'] as List<dynamic>).map((e) => (e as num).toInt()).toSet(),
+);
 
 Map<String, dynamic> _$SetSupportToJson(SetSupport instance) =>
-    <String, dynamic>{
-      'values': instance.values.toList(),
-    };
+    <String, dynamic>{'values': instance.values.toList()};
 ''')
 @JsonSerializable()
 class SetSupport {
@@ -172,58 +163,45 @@ class SetSupport {
   SetSupport(this.values);
 }
 
-@ShouldThrow(
-  '''
+@ShouldThrow('''
 Could not generate `toJson` code for `watch`.
 To support the type `Stopwatch` you can:
-$converterOrKeyInstructions''',
-)
+$converterOrKeyInstructions''')
 @JsonSerializable(createFactory: false)
 class NoSerializeFieldType {
   Stopwatch? watch;
 }
 
-@ShouldThrow(
-  '''
+@ShouldThrow('''
 Could not generate `fromJson` code for `watch`.
 To support the type `Stopwatch` you can:
-$converterOrKeyInstructions''',
-)
+$converterOrKeyInstructions''')
 @JsonSerializable(createToJson: false)
 class NoDeserializeFieldType {
   Stopwatch? watch;
 }
 
-@ShouldThrow(
-  '''
+@ShouldThrow('''
 Could not generate `toJson` code for `durationDateTimeMap` because of type `Duration`.
-Map keys must be one of: Object, dynamic, enum, String, BigInt, DateTime, int, Uri.''',
-)
+Map keys must be one of: Object, dynamic, enum, String, BigInt, DateTime, int, Uri.''')
 @JsonSerializable(createFactory: false)
 class NoSerializeBadKey {
   late Map<Duration, DateTime> durationDateTimeMap;
 }
 
-@ShouldThrow(
-  '''
+@ShouldThrow('''
 Could not generate `fromJson` code for `durationDateTimeMap` because of type `Duration`.
-Map keys must be one of: Object, dynamic, enum, String, BigInt, DateTime, int, Uri.''',
-)
+Map keys must be one of: Object, dynamic, enum, String, BigInt, DateTime, int, Uri.''')
 @JsonSerializable(createToJson: false)
 class NoDeserializeBadKey {
   late Map<Duration, DateTime> durationDateTimeMap;
 }
 
-@ShouldGenerate(
-  r'''
+@ShouldGenerate(r'''
 Map<String, dynamic> _$IncludeIfNullOverrideToJson(
-        IncludeIfNullOverride instance) =>
-    <String, dynamic>{
-      'number': instance.number,
-      if (instance.str case final value?) 'str': value,
-    };
-''',
-)
+  IncludeIfNullOverride instance,
+) => <String, dynamic>{'number': instance.number, 'str': ?instance.str};
+''')
 @JsonSerializable(createFactory: false, includeIfNull: false)
 class IncludeIfNullOverride {
   @JsonKey(includeIfNull: true)
@@ -232,9 +210,7 @@ class IncludeIfNullOverride {
 }
 
 // https://github.com/google/json_serializable.dart/issues/7 regression
-@ShouldThrow(
-  'The class `NoCtorClass` has no default constructor.',
-)
+@ShouldThrow('The class `NoCtorClass` has no default constructor.')
 @JsonSerializable()
 class NoCtorClass {
   late final int member;
@@ -302,7 +278,7 @@ class IgnoredFieldClass {
 @ShouldThrow(
   'Cannot populate the required constructor argument: ignoredTrueField. It is '
   'assigned to a field not meant to be used in fromJson.',
-  element: '',
+  element: 'new',
 )
 @JsonSerializable()
 class IgnoredFieldCtorClass {
@@ -339,7 +315,7 @@ class IgnoreAndIncludeFromJsonFieldCtorClass {
 @ShouldThrow(
   'Cannot populate the required constructor argument: '
   '_privateField. It is assigned to a private field.',
-  element: '',
+  element: 'new',
 )
 @JsonSerializable()
 class PrivateFieldCtorClass {
@@ -373,7 +349,7 @@ class JsonValueWithBool {
 
 enum BadEnum {
   @JsonValue(true)
-  value
+  value,
 }
 
 @ShouldGenerate(r'''const _$GoodEnumEnumMap = {
@@ -398,19 +374,20 @@ enum GoodEnum {
   @JsonValue(42)
   intValue,
   @JsonValue(null)
-  nullValue
+  nullValue,
 }
 
 @ShouldGenerate(r'''
 FieldWithFromJsonCtorAndTypeParams _$FieldWithFromJsonCtorAndTypeParamsFromJson(
-        Map<String, dynamic> json) =>
-    FieldWithFromJsonCtorAndTypeParams()
-      ..customOrders = json['customOrders'] == null
-          ? null
-          : MyList<GeneralTestClass2, int>.fromJson((json['customOrders']
-                  as List<dynamic>)
+  Map<String, dynamic> json,
+) => FieldWithFromJsonCtorAndTypeParams()
+  ..customOrders = json['customOrders'] == null
+      ? null
+      : MyList<GeneralTestClass2, int>.fromJson(
+          (json['customOrders'] as List<dynamic>)
               .map((e) => GeneralTestClass2.fromJson(e as Map<String, dynamic>))
-              .toList());
+              .toList(),
+        );
 ''')
 @JsonSerializable(createToJson: false)
 class FieldWithFromJsonCtorAndTypeParams {
@@ -447,17 +424,17 @@ mixin _PropInMixinI448RegressionMixin {
 
 @ShouldGenerate(r'''
 PropInMixinI448Regression _$PropInMixinI448RegressionFromJson(
-        Map<String, dynamic> json) =>
-    PropInMixinI448Regression()
-      ..nullable = (json['nullable'] as num).toInt()
-      ..notNullable = (json['notNullable'] as num).toInt();
+  Map<String, dynamic> json,
+) => PropInMixinI448Regression()
+  ..nullable = (json['nullable'] as num).toInt()
+  ..notNullable = (json['notNullable'] as num).toInt();
 
 Map<String, dynamic> _$PropInMixinI448RegressionToJson(
-        PropInMixinI448Regression instance) =>
-    <String, dynamic>{
-      'nullable': instance.nullable,
-      'notNullable': instance.notNullable,
-    };
+  PropInMixinI448Regression instance,
+) => <String, dynamic>{
+  'nullable': instance.nullable,
+  'notNullable': instance.notNullable,
+};
 ''')
 @JsonSerializable()
 class PropInMixinI448Regression with _PropInMixinI448RegressionMixin {
@@ -465,17 +442,13 @@ class PropInMixinI448Regression with _PropInMixinI448RegressionMixin {
   late int notNullable;
 }
 
-@ShouldGenerate(
-  r'''
+@ShouldGenerate(r'''
 IgnoreUnannotated _$IgnoreUnannotatedFromJson(Map<String, dynamic> json) =>
     IgnoreUnannotated()..annotated = (json['annotated'] as num).toInt();
 
 Map<String, dynamic> _$IgnoreUnannotatedToJson(IgnoreUnannotated instance) =>
-    <String, dynamic>{
-      'annotated': instance.annotated,
-    };
-''',
-)
+    <String, dynamic>{'annotated': instance.annotated};
+''')
 @JsonSerializable(ignoreUnannotated: true)
 class IgnoreUnannotated {
   @JsonKey()
@@ -484,17 +457,13 @@ class IgnoreUnannotated {
   late int unannotated;
 }
 
-@ShouldGenerate(
-  r'''
+@ShouldGenerate(r'''
 SubclassedJsonKey _$SubclassedJsonKeyFromJson(Map<String, dynamic> json) =>
     SubclassedJsonKey()..annotatedWithSubclass = (json['bob'] as num).toInt();
 
 Map<String, dynamic> _$SubclassedJsonKeyToJson(SubclassedJsonKey instance) =>
-    <String, dynamic>{
-      'bob': instance.annotatedWithSubclass,
-    };
-''',
-)
+    <String, dynamic>{'bob': instance.annotatedWithSubclass};
+''')
 @JsonSerializable(ignoreUnannotated: true)
 class SubclassedJsonKey {
   @MyJsonKey()
@@ -505,19 +474,15 @@ class MyJsonKey extends JsonKey {
   const MyJsonKey() : super(name: 'bob');
 }
 
-@ShouldGenerate(
-  r'''
+@ShouldGenerate(r'''
 OverrideGetterExampleI613 _$OverrideGetterExampleI613FromJson(
-        Map<String, dynamic> json) =>
-    OverrideGetterExampleI613()..id = json['id'] as String;
+  Map<String, dynamic> json,
+) => OverrideGetterExampleI613()..id = json['id'] as String;
 
 Map<String, dynamic> _$OverrideGetterExampleI613ToJson(
-        OverrideGetterExampleI613 instance) =>
-    <String, dynamic>{
-      'id': instance.id,
-    };
-''',
-)
+  OverrideGetterExampleI613 instance,
+) => <String, dynamic>{'id': instance.id};
+''')
 @JsonSerializable()
 class OverrideGetterExampleI613 extends OverrideGetterExampleI613Super {
   @override
@@ -584,11 +549,8 @@ class ExtraParamToJson {
 
 @ShouldGenerate(r'''
 Map<String, dynamic> _$Issue1038RegressionTestToJson(
-        Issue1038RegressionTest instance) =>
-    <String, dynamic>{
-      'id': instance.id,
-      'ean': instance.ean,
-    };
+  Issue1038RegressionTest instance,
+) => <String, dynamic>{'id': instance.id, 'ean': instance.ean};
 ''')
 @JsonSerializable(createFactory: false)
 class Issue1038RegressionTest {
@@ -600,16 +562,14 @@ class Issue1038RegressionTest {
   Issue1038RegressionTest.ean(this.ean) : id = null;
 }
 
-@ShouldGenerate(
-  r'''
+@ShouldGenerate(r'''
 TearOffFromJsonClass _$TearOffFromJsonClassFromJson(
-        Map<String, dynamic> json) =>
-    TearOffFromJsonClass(
-      TearOffValueClass(json['value'] as String),
-      TearOffValueClass.fromJson(json['factoryValue'] as String),
-    );
-''',
-)
+  Map<String, dynamic> json,
+) => TearOffFromJsonClass(
+  TearOffValueClass(json['value'] as String),
+  TearOffValueClass.fromJson(json['factoryValue'] as String),
+);
+''')
 @JsonSerializable(createToJson: false)
 class TearOffFromJsonClass {
   TearOffFromJsonClass(this.value, this.factoryValue);
@@ -626,4 +586,193 @@ class TearOffValueClass {
   factory TearOffValueClass.fromJson(String value) => TearOffValueClass(value);
 
   final String value;
+}
+
+@ShouldGenerate(r'''
+const _$JsonSchemaTestClassJsonSchema = {
+  r'$schema': 'https://json-schema.org/draft/2020-12/schema',
+  'type': 'object',
+  'properties': {
+    'name': {'type': 'string'},
+    'age': {'type': 'integer'},
+  },
+  'required': ['name', 'age'],
+};
+''', contains: true)
+@JsonSerializable(
+  createJsonSchema: true,
+  createFactory: false,
+  createToJson: false,
+)
+class JsonSchemaTestClass {
+  final String name;
+  final int age;
+
+  JsonSchemaTestClass(this.name, this.age);
+}
+
+@ShouldGenerate(r'''
+DateTimeUtcTestClass _$DateTimeUtcTestClassFromJson(
+  Map<String, dynamic> json,
+) => DateTimeUtcTestClass(DateTime.parse(json['date'] as String));
+
+Map<String, dynamic> _$DateTimeUtcTestClassToJson(
+  DateTimeUtcTestClass instance,
+) => <String, dynamic>{'date': instance.date.toUtc().toIso8601String()};
+''')
+@JsonSerializable(dateTimeUtc: true)
+class DateTimeUtcTestClass {
+  final DateTime date;
+
+  DateTimeUtcTestClass(this.date);
+}
+
+@ShouldThrow(
+  'Could not generate `fromJson` code for `record` because of type '
+  '`void Function()`.',
+)
+@JsonSerializable()
+class RecordWithFunction {
+  final (int, void Function()) record;
+
+  RecordWithFunction(this.record);
+}
+
+@ShouldThrow(
+  'Could not generate `fromJson` code for `record` because of type '
+  '`void Function()`.',
+)
+@JsonSerializable()
+class RecordWithNamedFunction {
+  final ({int a, void Function() b}) record;
+
+  RecordWithNamedFunction(this.record);
+}
+
+@ShouldThrow(
+  'Could not generate `fromJson` code for `record` because of type '
+  '`void Function()`.',
+)
+@JsonSerializable()
+class RecordWithSinglePositionalFunction {
+  final (void Function(),) record;
+
+  RecordWithSinglePositionalFunction(this.record);
+}
+
+class RecordConverter1 extends JsonConverter<(int, int), Map<String, dynamic>> {
+  const RecordConverter1();
+  @override
+  (int, int) fromJson(Map<String, dynamic> json) => (0, 0);
+  @override
+  Map<String, dynamic> toJson((int, int) object) => {};
+}
+
+class RecordConverter2 extends JsonConverter<(int, int), Map<String, dynamic>> {
+  const RecordConverter2();
+  @override
+  (int, int) fromJson(Map<String, dynamic> json) => (0, 0);
+  @override
+  Map<String, dynamic> toJson((int, int) object) => {};
+}
+
+@ShouldThrow('Found more than one matching converter for `(int, int)`.')
+@JsonSerializable()
+@RecordConverter1()
+@RecordConverter2()
+class RecordDoubleConverter {
+  late (int, int) record;
+}
+
+class SingleRecordConverter1
+    extends JsonConverter<(int,), Map<String, dynamic>> {
+  const SingleRecordConverter1();
+  @override
+  (int,) fromJson(Map<String, dynamic> json) => (0,);
+  @override
+  Map<String, dynamic> toJson((int,) object) => {};
+}
+
+class SingleRecordConverter2
+    extends JsonConverter<(int,), Map<String, dynamic>> {
+  const SingleRecordConverter2();
+  @override
+  (int,) fromJson(Map<String, dynamic> json) => (0,);
+  @override
+  Map<String, dynamic> toJson((int,) object) => {};
+}
+
+@ShouldThrow('Found more than one matching converter for `(int,)`.')
+@JsonSerializable()
+@SingleRecordConverter1()
+@SingleRecordConverter2()
+class RecordSingleDoubleConverter {
+  late (int,) record;
+}
+
+class NullableRecordConverter1
+    extends JsonConverter<(int, int)?, Map<String, dynamic>> {
+  const NullableRecordConverter1();
+  @override
+  (int, int)? fromJson(Map<String, dynamic> json) => null;
+  @override
+  Map<String, dynamic> toJson((int, int)? object) => {};
+}
+
+class NullableRecordConverter2
+    extends JsonConverter<(int, int)?, Map<String, dynamic>> {
+  const NullableRecordConverter2();
+  @override
+  (int, int)? fromJson(Map<String, dynamic> json) => null;
+  @override
+  Map<String, dynamic> toJson((int, int)? object) => {};
+}
+
+@ShouldThrow('Found more than one matching converter for `(int, int)?`.')
+@JsonSerializable()
+@NullableRecordConverter1()
+@NullableRecordConverter2()
+class RecordNullableDoubleConverter {
+  late (int, int)? record;
+}
+
+class NamedRecordConverter1
+    extends JsonConverter<({int a, int b}), Map<String, dynamic>> {
+  const NamedRecordConverter1();
+  @override
+  ({int a, int b}) fromJson(Map<String, dynamic> json) => (a: 0, b: 0);
+  @override
+  Map<String, dynamic> toJson(({int a, int b}) object) => {};
+}
+
+class NamedRecordConverter2
+    extends JsonConverter<({int a, int b}), Map<String, dynamic>> {
+  const NamedRecordConverter2();
+  @override
+  ({int a, int b}) fromJson(Map<String, dynamic> json) => (a: 0, b: 0);
+  @override
+  Map<String, dynamic> toJson(({int a, int b}) object) => {};
+}
+
+@ShouldThrow('Found more than one matching converter for `({int a, int b})`.')
+@JsonSerializable()
+@NamedRecordConverter1()
+@NamedRecordConverter2()
+class RecordNamedDoubleConverter {
+  late ({int a, int b}) record;
+}
+
+@ShouldThrow('''
+Could not generate `fromJson` code for `field` because of type `void Function()`.''')
+@JsonSerializable(createToJson: false)
+class UnsupportedNestedFunctionType {
+  late List<void Function()> field;
+}
+
+@ShouldThrow('''
+Could not generate `fromJson` code for `map` because of type `(int, int)`.
+Map keys must be one of: Object, dynamic, enum, String, BigInt, DateTime, int, Uri.''')
+@JsonSerializable(createToJson: false)
+class UnsupportedMapKeyRecord {
+  late Map<(int, int), String> map;
 }